backend/requirements.txt
asgiref==3.9.1
boto3==1.40.27
botocore==1.40.27
Django==5.2.6
django-appconf==1.1.0
django-cors-headers==4.8.0
django-fsm==3.0.0
django-fsm-2==4.0.0
django-fsm-admin2==0.1.3
django-fsm-log==3.1.0
django-mongodb-backend==5.2.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
jmespath==1.0.1
joblib==1.5.2
mixes==1.0
numpy==2.3.3
PyJWT==2.10.1
pymongo==4.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
urllib3==2.5.0

django-filter~=25.1

pillow==11.3.0
